Convert TagHelperFactsService to an interface
