+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/Razor/ComponentAccessibilityCodeActionProviderTest.cs
@@ -48,7 +48,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(0, 1));
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -78,7 +78,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(0, 0));
         context.CodeDocument.SetFileKind(FileKinds.Legacy);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -107,7 +107,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9));
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -136,7 +136,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9), supportsFileCreation: true);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -185,7 +185,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("GenericComponent", StringComparison.Ordinal), "GenericComponent".Length), supportsFileCreation: true);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -234,7 +234,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9), supportsFileCreation: true);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -266,7 +266,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9), supportsFileCreation: true);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -298,7 +298,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9), supportsFileCreation: false);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
@@ -328,7 +328,7 @@ public class ComponentAccessibilityCodeActionProviderTest : LanguageServerTestBa
         var location = new SourceLocation(cursorPosition, -1, -1);
         var context = CreateRazorCodeActionContext(request, location, documentPath, contents, new SourceSpan(contents.IndexOf("Component", StringComparison.Ordinal), 9), supportsFileCreation: false);
 
-        var provider = new ComponentAccessibilityCodeActionProvider(new DefaultTagHelperFactsService());
+        var provider = new ComponentAccessibilityCodeActionProvider(new TagHelperFactsService());
 
         // Act
         var commandOrCodeActionContainer = await provider.ProvideAsync(context, default);
